ranged enemy
the arrows dealing damage was occasionally being called twice, doing
twice the dame.  adjusted code to do appropriate amount of damage

+++ Assets/Scripts/Enemy/EnemyHealth.cs
@@ -2,8 +2,8 @@
 using UnityEngine.UI;
 
 public class EnemyHealth : MonoBehaviour {
-    public int startingHealth = 100;            // The amount of health the enemy starts the game with.
-    public int currentHealth;                   // The current health the enemy has.
+    public float startingHealth = 100;            // The amount of health the enemy starts the game with.
+    public float currentHealth;                   // The current health the enemy has.
     public float lookRange = 20.0f;
     public float attackRange = 5.0f;
     public float attackCooldownTime = 1.0f;
@@ -20,7 +20,7 @@ public class EnemyHealth : MonoBehaviour {
         isRanged = GetComponent<EnemyArrowShootingController>() != null;
     }
 
-    public void TakeDamage(int amount)
+    public void TakeDamage(float amount)
     {
         currentHealth -= amount;
